+++ R/utils.R
@@ -622,3 +622,15 @@ sub_name_space_accessor_function = function(self, name) {
 "%in_list%" = \(lhs_element, rhs_list) rhs_list |>
   sapply("==", lhs_element) |>
   any()
+
+
+# helper used to validate inputs passed to pl$set_options()
+is_bool = function(x) {
+  is.logical(x) && length(x) == 1 && !is.na(x)
+}
+
+# takes a list of dtypes (for example from $schema), returns a named vector
+# indicating which are Structs
+dtypes_are_struct = function(dtypes) {
+  sapply(dtypes, \(dt) pl$same_outer_dt(dt, pl$Struct()))
+}