+++ src/rust/src/rdataframe/mod.rs
@@ -299,26 +299,8 @@ impl DataFrame {
         s.into_series().into()
     }
 
-    pub fn unnest(&self, names: Nullable<Vec<String>>) -> List {
-        let names = if let Some(vec_string) = null_to_opt(names) {
-            vec_string
-        } else {
-            //missing names choose to unnest any column of DataType Struct
-            self.0
-                .dtypes()
-                .iter()
-                .zip(self.0.get_column_names().iter())
-                .filter(|(dtype, _)| matches!(dtype, pl::DataType::Struct(_)))
-                .map(|(_, y)| y.to_string())
-                .collect::<Vec<String>>()
-        };
-
-        r_result_list(
-            self.0
-                .unnest(names)
-                .map(DataFrame)
-                .map_err(|err| format!("in unnest: {:?}", err)),
-        )
+    pub fn unnest(&self, names: Vec<String>) -> RResult<Self> {
+        self.lazy().unnest(names)?.collect()
     }
 
     pub fn export_stream(&self, stream_ptr: &str) {
@@ -357,7 +339,7 @@ impl DataFrame {
         value_vars: Robj,
         value_name: Robj,
         variable_name: Robj,
-    ) -> Result<Self, String> {
+    ) -> RResult<Self> {
         let args = MeltArgs {
             id_vars: strings_to_smartstrings(robj_to!(Vec, String, id_vars)?),
             value_vars: strings_to_smartstrings(robj_to!(Vec, String, value_vars)?),
@@ -365,8 +347,11 @@ impl DataFrame {
             variable_name: robj_to!(Option, String, variable_name)?.map(|s| s.into()),
             streamable: false,
         };
-        let df = self.0.melt2(args).map_err(|s| s.to_string())?;
-        Ok(DataFrame(df))
+
+        self.0
+            .melt2(args)
+            .map_err(polars_to_rpolars_err)
+            .map(DataFrame)
     }
 
     pub fn pivot_expr(
@@ -378,7 +363,7 @@ impl DataFrame {
         sort_columns: Robj,
         aggregate_expr: Robj,
         separator: Robj,
-    ) -> Result<Self, String> {
+    ) -> RResult<Self> {
         let fun = if robj_to!(bool, maintain_order)? {
             pivot_stable
         } else {
@@ -394,11 +379,48 @@ impl DataFrame {
             robj_to!(Option, PLExpr, aggregate_expr)?,
             robj_to!(Option, str, separator)?,
         )
-        .map_err(|err| err.to_string())
-        .map(|ok| ok.into())
+        .map_err(polars_to_rpolars_err)
+        .map(DataFrame)
+    }
+
+    pub fn sample_n(
+        &self,
+        n: Robj,
+        with_replacement: Robj,
+        shuffle: Robj,
+        seed: Robj,
+    ) -> RResult<Self> {
+        self.0
+            .clone()
+            .sample_n(
+                robj_to!(usize, n)?,
+                robj_to!(bool, with_replacement)?,
+                robj_to!(bool, shuffle)?,
+                robj_to!(Option, u64, seed)?,
+            )
+            .map_err(polars_to_rpolars_err)
+            .map(DataFrame)
+    }
+
+    pub fn sample_frac(
+        &self,
+        frac: Robj,
+        with_replacement: Robj,
+        shuffle: Robj,
+        seed: Robj,
+    ) -> RResult<Self> {
+        self.0
+            .clone()
+            .sample_frac(
+                robj_to!(f64, frac)?,
+                robj_to!(bool, with_replacement)?,
+                robj_to!(bool, shuffle)?,
+                robj_to!(Option, u64, seed)?,
+            )
+            .map_err(polars_to_rpolars_err)
+            .map(DataFrame)
     }
 }
-use crate::utils::wrappers::null_to_opt;
 impl DataFrame {
     pub fn to_list_result(&self) -> Result<Robj, pl::PolarsError> {
         //convert DataFrame to Result of to R vectors, error if DataType is not supported