+++ NEWS.md
@@ -1,11 +1,31 @@
 # polars (development version)
 
+## Breaking changes
+
+- Setting and getting polars options is now made with `pl$options`, 
+  `pl$set_options()` and `pl$reset_options()` (#384).
+- Method `$with_column()` has been removed (it was deprecated since 0.8.0). Use 
+  `$with_columns()` instead (#402).
+- Subnamespace `$arr` has been removed (it was deprecated since 0.8.1). Use `$list`
+  instead (#402).
+
+## What's changed
+
+- New method `$unnest()` for `LazyFrame` (#397).
+- New method `$sample()` for `DataFrame` (#399).
+- New method `$meta$tree_format()` to display an `Expr` as a tree (#401).
+
+# polars 0.8.1
+
 ## What's changed
 
 - New string method `to_titlecase()` (#371).
+- Although stated in news for PR (#334) `strip = true` was not actually set for the 
+  "release-optimized" compilation profile. Now it is, but the binary sizes seems unchanged (#377).
 - New vignette on best practices to improve `polars` performance (#188).
-- Subnamespace name "arr" as in  `<Expr>$arr$` & `<Series>$arr$` is deprecated in favor of "list". 
-  Finally at polars 0.9.0 the "arr" will be removed (#375).
+- Subnamespace name "arr" as in  `<Expr>$arr$` & `<Series>$arr$` is deprecated 
+  in favor of "list". The subnamespace "arr" will be removed in polars 0.9.0 (#375).
+
 # polars 0.8.0
 
 ## CHANGES DUE TO RUST-POLARS 0.32.0